myshopfurniture-docker-master/requirements.txt
Django==1.9.8
django-appconf==1.0.1
django-carton==1.2.1
django-filer==1.2.5
django-filter==1.0.1
django-image-cropping==1.0.4
django-mptt==0.8.6
django-nested-admin==3.0.12
django-nested-inline==0.3.6
django-polymorphic==0.8.1
django-taggit==0.21.3
django-tinymce==2.4.0
easy-thumbnails==2.3
MySQL-python==1.2.5
Pillow==3.4.2
python-monkey-business==1.0.0
six==1.10.0
Unidecode==0.4.19